Rename the util module to file_util.
