+++ lindera-core/src/lib.rs
@@ -2,10 +2,10 @@ pub mod character_definition;
 pub mod connection;
 pub mod dictionary_builder;
 pub mod error;
+pub mod file_util;
 pub mod prefix_dict;
 pub mod unknown_dictionary;
 pub mod user_dictionary;
-pub mod util;
 pub mod viterbi;
 pub mod word_entry;
